# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations REST API",
        "description": "API справочника организаций",
        "version": "1.0.0"
    },
    "paths": {
        "/api/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список организаций с фильтрацией",
                "operationId": "42efb13746ba71d73c330585f09ecef8",
                "parameters": [
                    {
                        "name": "filter[activity_id]",
                        "in": "query",
                        "description": "Фильтр по ID вида деятельности",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "filter[building_id]",
                        "in": "query",
                        "description": "Фильтр по ID здания",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "filter[bbox][latitude]",
                        "in": "query",
                        "description": "Центр поиска — широта",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.751244
                        }
                    },
                    {
                        "name": "filter[bbox][longitude]",
                        "in": "query",
                        "description": "Центр поиска — долгота",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 37.618423
                        }
                    },
                    {
                        "name": "filter[bbox][radius]",
                        "in": "query",
                        "description": "Радиус поиска (в км)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 5
                        }
                    },
                    {
                        "name": "filter[bbox][min_lat]",
                        "in": "query",
                        "description": "Минимальная широта (для bounding box)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.7
                        }
                    },
                    {
                        "name": "filter[bbox][max_lat]",
                        "in": "query",
                        "description": "Максимальная широта (для bounding box)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.8
                        }
                    },
                    {
                        "name": "filter[bbox][min_lng]",
                        "in": "query",
                        "description": "Минимальная долгота (для bounding box)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 37.6
                        }
                    },
                    {
                        "name": "filter[bbox][max_lng]",
                        "in": "query",
                        "description": "Максимальная долгота (для bounding box)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 37.7
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Поиск по названию организации",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "68e75650b204083886e125e02aaa2f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Organization": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "building_id": {
                        "type": "integer"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ],
    "tags": [
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}